add /{email} REST methods

+++ friendsify-person-service/src/main/java/aau/at/friendsifypersonservice/controllers/PersonController.java
@@ -31,6 +31,16 @@ public class PersonController {
         return ResponseEntity.ok(p);
     }
 
+    @GetMapping("/{email}")
+    public ResponseEntity<Person> getPersonByEmail(@PathVariable("email")String email) throws PersonNotFoundException {
+
+        Person p = personDao.findByEmail(email);
+        if (p == null){
+            throw new PersonNotFoundException("Person not found by email: "+email);
+        }
+        return ResponseEntity.ok(p);
+    }
+
     @PostMapping()
     public Person createPerson(@RequestBody Person person){
         return personDao.save(person);
@@ -47,12 +57,35 @@ public class PersonController {
         return ResponseEntity.ok(updatedPerson);
     }
 
+    @PutMapping("/{email}")
+    public ResponseEntity<Person> updatePerson(@PathVariable("email") String email, @RequestBody Person newPerson) throws PersonNotFoundException {
+        Person storedPerson = personDao.findByEmail(email);
+        if (storedPerson==null) throw new PersonNotFoundException("Person not found by email: "+email);
+
+        storedPerson.update(newPerson);
+        Person updatedPerson= personDao.save(storedPerson);
+        return ResponseEntity.ok(updatedPerson);
+    }
+
     @DeleteMapping("/{id}")
     public Map<String, Boolean> deletePerson(@PathVariable("id") Long id)
             throws PersonNotFoundException {
 
         Person storedPerson = personDao.findById(id)
                 .orElseThrow(() -> new PersonNotFoundException("Person not found by id: "+id));
+
+        personDao.delete(storedPerson);
+        Map<String, Boolean> response = new HashMap<>();
+        response.put("deleted", Boolean.TRUE);
+        return response;
+    }
+
+    @DeleteMapping("/{email}")
+    public Map<String, Boolean> deletePerson(@PathVariable("email") String email)
+            throws PersonNotFoundException {
+
+        Person storedPerson = personDao.findByEmail(email);
+        if (storedPerson==null) throw new PersonNotFoundException("Person not found by email: "+email);
 
         personDao.delete(storedPerson);
         Map<String, Boolean> response = new HashMap<>();
